New [TypeOps] module.
The desugaring pass introduces a truckload of intermediate names that are often
redundant with user-provided names. This new optimization performs a symbolic
evaluation of equality-relations in the arguments of a function, and removes
redundant quantifiers. It also removes redundant equality permissions that were
present in the return type of a function.

This turns the following type:

∀(a :: ∗).
∀(b :: ∗).
  ∀(f :: term).
    ∀(xs :: term).
      ∀(/c0 :: term).
        ∀(/root1 :: term).
          (=/root1 | /root1 @ ((=f | f @ ∀(x :: term).
            ∀(/c2 :: term).
              ∀(/root3 :: term). (=/root3 | /root3 = /c2 ∗ /c2 @ (=x | x @ a)) -> (b | /root3 = /c2)), =/c0) ∗ /c0 @ (=xs | xs @ list a)) ->
          (list b | /root1 @ ((=f | f @ ∀(x :: term).
            ∀(/c4 :: term).
              ∀(/root5 :: term). (=/root5 | /root5 = /c4 ∗ /c4 @ (=x | x @ a)) -> (b | /root5 = /c4)), =/c0))

into:

∀(a :: ∗).
  ∀(b :: ∗).
    ∀(f :: term).
      ∀(xs :: term).
        ∀(/root1 :: term).
          (=/root1 | xs @ list a ∗ f @ ∀(x :: term). (=x | x @ a) -> b ∗ /root1 @ (=f, =xs)) ->
          (list b | /root1 @ (=f, =xs) ∗ f @ ∀(x :: term). (=x | x @ a) -> b)

One optimisation left that we still should perform is remove duplicable
permissions in the right-hand side of a function that are also in the left-hand
side.

Both these optimizations require a complete environment to run, so they're
performed on-the-fly while type-checking, rather than in a separate pass.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/hamlet@375 0f7ccc15-3e29-4385-878f-320e1e960475