Fix the \forall vs \forall merge.
This makes the [map] function type-check. It is strange that we should go
through this case and not through the fast-path. This happens because a type
variable that's local to a branch has been unified with, say, int. However, the
[equal] function can't understand this type in the top-level environment because
the unified variable is now a union-find link, which is only valid in the local
environment.

One solution would be to call [PersistentUnionFind.repr] on all points with
their local environment, and then run them through [equal]. We'll keep that for
later, because a better solution follows.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/hamlet@384 0f7ccc15-3e29-4385-878f-320e1e960475