More work on contraints/patterns/concrete types.
- More bugs uncovered: we previously forgot to expand nominal type [foo] into
[Foo] when [foo] only had one branch, this is now performed.
- We did *not* propagate a type annotation's field components down into
the fields of a [EConstruct] expression. This is now performed.
- If, when merging, the destination point had a concrete type annotation, we
would just use it as the resulting type. We now dive into the concrete type
annotation and recursively merge its fields, which is good in case the user only
provided a type annotation for one field only: we are able to extract
information from the other fields.
- We still need to merge (as mentioned in the previous commit message) type
annotation.
- A bug was uncovered in the desugaring phase (desugaring of type annotations in
patterns of match expressions).

Only 1. and 2. are fixed. 3. should be fixed but is not, so more debugging is
required. 4. and 5. are slated for later.