Interfaces: check facts only for "abstract" types.
- Some types can be declared with the "abstract" keyword (unrelated to
interfaces/signatures -- yet). For these we specify the facts. We only check
that the facts in the signature match those in the implementation for these
"abstract" types. For other types, since we don't have real abstract types, we
have the definition in the interface, which means we are able to properly infer
the fact.
- For these "abstract" types, we cannot declare the variance of parameters, so
we don't check variance at all when checking an implementation against an
interface.
- The syntax/treatment/desugaring of facts on "abstract" types, when those fact
refer to the type variables, doesn't work AT ALL.

All of this ought to be fixed when we implement real abstract types.