Add a bit of sophisticated logic in [Permissions].
- When performing "(t1 | p1) - (t2 | p2)", it looks as if this is simple.
However, in the presence of flexible variables, the order in which we perform
"t1 - t2", "add p1", "add p2" matters a lot, because if we pick the wrong order,
we may end up being stuck.
- It may look like an excessively complicated case. It's not! This is
super-important. Indeed, if "f" expects a function type "t", and we do a call
such as "f g", we need to decide whether the type of "g" is compatible with "t".
Syntactic equality (modulo alpha-conversion) only works in a limited set of
situations... as soon as, for instance, a conjunction is not in the same order,
syntactic equality fails.
  Therefore, we need to decide whether a function type is a subtype of another.
Function types contains bar types everywhere, so we need to able to reason
efficiently about these, hence this complicated logic.
  I had already encountered this problem for the series of [map] examples, but I
had not thought of this general search procedure until now, when trying to make
Gabriel Dos Reis' example work.
- Now we have 100 tests, and the hundredth test passes!