Parser changes.
The prefixed operators were too high so "!x + 1" wasn't parsed at all. I moved
the prefixed operators closer to the atomics but then I ran again in the problem
of 1 - 2 begin possibly 1 (-2). I split the atomics into two categories.

I think that's the right way to solve the problem but comments are welcome!