Tweak list-map3 so that it doesn't work anymore.
- When substituting "(t, u)" for "a" in a function type "[a] (...)", the
property that all types appearing in the function type are unfolded no longer
holds (which causes problems later on).
- I could write a pass that expands such function types, but I feel like the
problem is more general and it may be worth trying to tackle, say, a subtraction
such as "(a, a) - (=x, =y) * x @ a * y @ a".
- I wrote a pass that made sure absolutely everything was expanded in function
types; the consequence was that function types became much more heavyweight.
This might allow us to get rid of the extra, non user-provided names.