Get rid of the [p=p] instantiation in simpler_iter.
I implemented a very silly strategy that basically says:
when doing (t | p) - (u | q), consider the type variables in p and q. If there's
exactly one on each side, then if one of them is flexible, it should instantiate
to the other one. If there's one variable on one side and zero variable on the
other side, and the variable is flexible, it should instantiate to empty (which
is very dumb, because we could want to try to instantiate it to something
else, i.e. the remaining permissions that we haven't been able to extract on
the other side).