Fix the function comparison.
The function comparison was previously not correct; since it would speculatively
add/remove permissions to compare function types, the resulting environment
had leftover permissions that were not meant to remain in the environment after
the function comparison (this took me a while to understand!).

The current solution is correct, but it heavyweight, as it implies saving an
entire set of permissions and then adding it back into the environment. We can't
just throw away the temporary environment that was used to compare function
types because it often instantiates flexible variables as it goes.

I'm committing this in a branch waiting for a better solution to come up (new
attempt).