Try to strengthen the introduction of flexible/rigid variables.
In particular, try to associate the correct facts. Fix, while we're at it,
removing a floating permission from the environment in the duplicable case.
Some ugly code appears in FactInference, but it also makes the results of the
fact inference procedure more accurate...

The list module is still broken by a fairly subtle failure...