Several bugfixes in [Permissions].
- When comparing t τ and t τ' (nominal type application), we perform a two-way
subtraction, that is, τ - τ' followed by τ' - τ. If that works, we declare these
two type applications to be equal.
  This is important because there may be flexible variables hidden in depth that
need to be instantiated, and the two-way subtraction takes care of that.
- Make sure that when subtracting a floating permission, we do not just rely on
syntactic equality (that was fine when we just had variables), but perform a
full-fledged subtraction (so that "answer t* - answer t" with "t*" flexible
works).

This should also allow us to get ride of a few type annotations (commit
pending) in the hash table module.