TyImply is now permitted only on top of arrows, and encoded into TyAnd during the translation of surface syntax to core syntax.
There is in fact one place where we used implication on top of
a non-arrow, and that was in reflection.mzi, where Duplicable
has type [a] duplicable a => duplicability a. For the moment,
this is commented out.

I had to add a rule And-L in Permissions.sub_type. The code is
the same as the former Imply-R. I don't know if it is correct
(how long does the assumption last?). Anyway, all tests pass,
including some tests that used to fail, such as tyand05 :-)