Type abbreviations [part 2].
- Change the internal representation of type groups. A group of type definitions
is a record with two fields; one tells whether the bindings are mutually
recursive, the other gives the list of type definitions.
- Data types are mutually recursive, abstract type definitions are not (but it
doesn't matter), and type abbreviations are not recursive for the moment.
- The translation and kind-checking phases have been adapted to deal with this
change.
- The variance analysis seems to be in place for type abbreviations.

The substitution functions in [DataTypeGroup] still don't seem to work properly,
though, so more work is needed.