Found a new bug in [Merge].
While re-reading the code of the [Merge] operation, I found out that one place
in the code was overly optimistic. A few seconds later, we have a new failing
test, "merge20.mz".

This bug is actually related to the merge of type abbreviations, which is
currently not working for the same reasons that merge20.mz is failing.