Make assertions and type annotations sealing.
This is a change of semantics. Previously, when writing:

  « let x: mlist a = MCons { ... } in e »

the expression [e] would have [x @ Cons ...] available, meaning that a type
annotation was merely a check that the annotation could be satisfied, without
any effect whatsoever. This behavior has now changed: the example above now
yields [x @ mlist a], meaning we *lose information*.

Assertions behave similarly:

  « let x = MCons { ... } in assert x @ mlist a; x »

has type [mlist a], not [MCons].

This behavior is slightly different in the case of immutable data types; since
the type-checker can always save a copy of the original permission, folding it
to [list a] and then adding it back has no effect.