A variant of [fpiterator], replacing the [borrowed] idiom by a [focused] idiom
The idea of the [focused] idiom is to store the [rewind] function with
the value needed to rewind. This is heavier: we allocate more
functions that all have the implementation () -> ().

However, that solves the problem of the typechecker being confused by
the same function [surrender] having multiple very different types.

We are consequently able to write the [equal] function, and also the
(previously implemented) [identity], [filter] and [map] functions.

Note that (for unclear reasons, at least for me) the whole source is
very slow to typecheck: 17 seconds on my computer.