Properly implement lazy concat.
I decided to save time by using a ready-made implementation for lazy lists, from
the respected and well-known "OCaml Batteries Included" project. What a mistake!
The documentation stated that the [concat] function that performs a lazy
concatenation of lazy lists of lazy lists but the function actually is *NOT*
lazy, which resulted in the code still enumerating all the possibilities.