First attempt of interface for tubes allowing to send elements and have them back.
A tube is represented as it two endpoints. An endpoint is either a
sending endpoint or a receiving endpoint, and each endpoint follow a
protocol with two states: ready and standby. The sender endpoint is in
ready state when it is possible to send an element, and in standby
state when it waits for the permission on the element to come back.

Symmetrically, when a receiver endpoint is ready, it can get an
element from the channel, and in standby state it waits for the user
to give back the permission on it.

The standby states are thus parametrized by the element that have been
sent/received.