The big constructor rewrite, part 1.
This is the first batch of changes on the path to the removal of
FieldPermission. (Making field names behave as name-intro's in data type
definitions should be a piece of cake after that.)

- I made the mutability per-branch in the internal representation. This means
that whether a branch is mutable or immutable is attached to the branch itself,
not its parent data type.
- This allowed me to make sure branch _definitions_ and branch _types_ use the
same record ("branch"). This will allow us to ultimately put _types_ for branch
definitions, thus allowing permissions, name introductions, etc.
- What about being resolved/unresolved? A "branch" always contains a
resolved_datacon. When the data type definition is translated into the internal
syntax, initially, the "parent" data type is "TyUnknown" (because no rigid
variable has been introduced for the type). Once a type is bound in the
environment, the "TyUnknown" is replaced by the right "TyPoint" (when doing
"set_definition"). This means that the definitions that are stored in the
environment contain already-resolved branches! This simplifies things quite a
big.

Some remarks.
- I wasted quite some time because [Permissions] contained "method
resolved_branch" in a derived class. Since the parent class renamed this method
"branch", the override was no longer in effect. I changed all these cases to
"method!" (the equivalent of "@Overrides" in Java) to get compilation warnings
in the future.